Make accuracy_score and score return float explicitly

+++ sklearn/decomposition/_pca.py
@@ -708,7 +708,7 @@ class PCA(_BasePCA):
             Average log-likelihood of the samples under the current model.
         """
         xp, _ = get_namespace(X)
-        return xp.mean(self.score_samples(X))
+        return float(xp.mean(self.score_samples(X)))
 
     def _more_tags(self):
         return {"preserves_dtype": [np.float64, np.float32], "array_api_support": True}
